refactor: move backend logic to services folder
