+++ frontend/src/App.tsx
@@ -5,15 +5,9 @@ import { useUserContext } from './hooks/useUserContext';
 const App = () => {
   const { isLogged } = useUserContext();
 
-  return (
-    <>
-      {
-        isLogged
-          ? <Home />
-          : <Login />
-      }
-    </>
-  )
+  return isLogged
+    ? <Home />
+    : <Login />
 }
 
 export default App
